+++ resources/views/bootstrap-4/components/modals/modal.blade.php
@@ -5,16 +5,22 @@
                 @isset ($header)
                     <div {{ $header->attributes->class(['modal-header']) }}>
                         {{ $header }}
-                @else
-                    <div class="modal-header">
-                        <h1 class="modal-title" id="{{ $titleLabel }}">{{ $title }}</h1>
-                @endif
                         @if ($dismissable)
                             <button type="button" class="close" data-dismiss="modal" aria-label="{!! trans('blade-ui-kit-bootstrap::modal.close') !!}">
                                 <span aria-hidden="true">&times;</span>
                             </button>
                         @endif
                     </div>
+                @else
+                    <div class="modal-header">
+                        <h1 class="modal-title" id="{{ $titleLabel }}">{{ $title }}</h1>
+                        @if ($dismissable)
+                            <button type="button" class="close" data-dismiss="modal" aria-label="{!! trans('blade-ui-kit-bootstrap::modal.close') !!}">
+                                <span aria-hidden="true">&times;</span>
+                            </button>
+                        @endif
+                    </div>
+                @endisset
                     <div class="modal-body">
                         {!! $slot !!}
                     </div>