+++ resources/views/bootstrap-5/components/forms/inputs/checkbox.blade.php
@@ -4,9 +4,9 @@
         type="checkbox"
         id="{{ $id }}"
         value="{{ $value }}"
-        @if($isChecked)checked@endif
-        @if($hasErrors)aria-describedby="validation-{{ $name }}-feedback"@endif
-        class="form-check-input @if($hasErrors) is-invalid @endif"
+        {{ $isChecked ? 'checked' : '' }}
+        {{ $hasErrors ? 'aria-describedby="validation-' . $name . '-feedback"' : '' }}
+        class="form-check-input{{ $hasErrors ? ' is-invalid' : '' }}"
     />
     <label class="form-check-label" for="{{ $id }}">
         {!! $label !!}