+++ resources/views/bootstrap-5/components/modals/form.blade.php
@@ -6,14 +6,18 @@
                     @isset ($header)
                         <div {{ $header->attributes->class(['modal-header']) }}>
                             {{ $header }}
-                    @else
-                        <div class="modal-header">
-                            <h1 class="modal-title" id="{{ $titleLabel }}">{{ $title }}</h1>
-                    @endif
                             @if ($dismissable)
                                 <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="{!! trans('blade-ui-kit-bootstrap::modal.close') !!}"></button>
                             @endif
                         </div>
+                    @else
+                        <div class="modal-header">
+                            <h1 class="modal-title" id="{{ $titleLabel }}">{{ $title }}</h1>
+                            @if ($dismissable)
+                                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="{!! trans('blade-ui-kit-bootstrap::modal.close') !!}"></button>
+                            @endif
+                        </div>
+                    @endisset
                         <div class="modal-body">
                             {!! $slot !!}
                         </div>